requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.1
gspread==3.2.0
httplib2==0.17.0
idna==2.8
idna-ssl==1.1.0
multidict==4.7.4
numpy==1.18.1
oauth2client==4.1.3
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.13
typing-extensions==3.7.4.1
urllib3==1.25.8
websockets==6.0
yarl==1.4.2
